mod chacheslogic 別途呼び出しではなく、FileContainerの機能として管理するように変更 特定のフォルダ配下に格納すると公開されるようになる
mod FileContainer add events
onChangeでファイルの変更を通知できるようにした

ファイルでパス表現可能に変更

+++ src/editor/js/model/FileContainer.js
@@ -3,6 +3,7 @@ Copyright 2017 - 2018
 ***********************************************/
 /* v1.0.0 */
 import FileData from './FileData.js'
+import {EventEmitter} from 'events'
 
 /* ------------------------------------------------
   ファイル管理クラス FileContainer
@@ -22,6 +23,11 @@ export class FileContainer {
       description: ''
     }
     this.fileObjects = {}
+    this.ev = new EventEmitter ()
+  }
+
+  onChangeFiles (callback) {
+    this.ev.on('change', callback)
   }
 
   setId (id) {
@@ -40,12 +46,18 @@ export class FileContainer {
     return this.container.projectName
   }
 
-  getFiles () {
+  getFiles ( parentPath = null) {
     // 配列のキーを取り出す
-    var ret = []
+    let ret = []
     for (var key in this.container.files) {
       if (!this.container.files[key].truncated) {
-        ret.push(key)
+        if (!parentPath) {
+          ret.push(key)
+        } else {
+          if(key.indexOf(parentPath)==0 && key.indexOf('/', parentPath.length)==-1 ) {
+            ret.push(key)
+          }
+        }
       }
     }
     return ret
@@ -68,24 +80,34 @@ export class FileContainer {
   }
 
   putFile (file) {
-    var filename = file.getFilename()
+    const filename = file.getFilename()
     this.container.files[filename] = file.getFileData()
     this.container.lastUpdatedTime = new Date().getTime()
+    this.ev.emit('change', filename)
     return true
   }
 
   renameFile (filename, newName) {
-    var file = this.getFile(filename)
-    file.setFilename(newName)
-    this.putFile(file)
-    delete this.container.files[filename]
+    if (filename in this.container.files) {
+      let file = new FileData(this.container.files[filename])
+      file.setFilename(newName)
+      delete this.container.files[filename]
+      this.putFile(file)
+      this.container.lastUpdatedTime = new Date().getTime()
+      return true
+    }
+    return false
   }
 
   removeFile (filename) {
-    var file = this.getFile(filename)
-    file.remove()
-    this.putFile(file)
-    this.container.lastUpdatedTime = new Date().getTime()
+    if (filename in this.container.files) {
+      let file = new FileData(this.container.files[filename])
+      file.remove()
+      this.putFile(file)
+      this.container.lastUpdatedTime = new Date().getTime()
+      return true
+    }
+    return false
   }
 
   init () {
@@ -122,6 +144,7 @@ export class FileContainer {
   setContainer (container) {
     this.container = container
     this.fileObjects = {}
+    this.ev.emit('change', null)
   }
 
   getContainer () {
@@ -138,7 +161,7 @@ export class FileContainer {
   }
 
   getGistData () {
-    var gistdata = {
+    let gistdata = {
       description: this.container.projectName + '\n' + this.container.description,
       public: this.container.public,
       files: this.container.files
